py/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "imbrace"
version = "1.0.0"
description = "Official Python SDK for Imbrace — dual auth (API Key + Access Token)"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
authors = [
  { name = "Imbrace Team", email = "dev@imbrace.co" }
]
keywords = ["imbrace", "sdk", "api", "automation"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: MIT License",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "httpx>=0.27.0",
  "pydantic>=2.0.0",
  "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://imbrace.co"
Repository = "https://github.com/imbrace/sdk"
Documentation = "https://docs.imbrace.co/sdk/python"

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-httpx>=0.30.0",
  "ruff>=0.4.0",
  "mypy>=1.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/imbrace"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests/unit", "tests/integration"]
markers = ["integration: marks tests as integration tests (require live API key)"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
exclude = ["site-packages", "\\.venv", "tests/"]
